requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.11.0
python-dotenv>=1.0.0
httpx>=0.24.0
paho-mqtt>=2.1.0
js2py>=0.74
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.2.0
redis==5.0.1
pydantic_settings>=0.2.0
requests>=2.31.0
pyotp==2.9.0
qrcode>=7.3.1
pillow>=9.0.0